ISelector: Coordinates stored as doubles

+++ src/auth_environment/backend/ISelector.java
@@ -4,14 +4,14 @@ public interface ISelector {
 	
 	public void chooseElement(int index);
 	
-	public void chooseCoordinates(int x, int y); 
+	public void chooseCoordinates(double x, double y); 
 	
 	// Returns index of the selected GameElement. 
 	public int getElementIndex();
 	
 	// Returns X and Y coordinates of the selected Map Tile. 
-	public int getX();
-	public int getY(); 
+	public double getX();
+	public double getY(); 
 	
 	// For debugging and testing
 	public void printIndex();
